+++ mayan/apps/document_indexing/models.py
@@ -381,8 +381,8 @@ class IndexInstanceNode(MPTTModel):
     objects = IndexInstanceNodeManager()
 
     class Meta:
-        verbose_name = _('Index node instance')
-        verbose_name_plural = _('Indexes node instances')
+        verbose_name = _('Index instance node')
+        verbose_name_plural = _('Indexes instances node')
 
     def __str__(self):
         return self.value
@@ -443,6 +443,10 @@ class IndexInstanceNode(MPTTModel):
                 result.append(force_text(node))
 
         return ' / '.join(result)
+    get_full_path.help_text = _(
+        'The path to the cabinet including all ancestors.'
+    )
+    get_full_path.short_description = _('Full path')
 
     def get_item_count(self, user):
         if self.index_template_node.link_documents:
@@ -498,3 +502,10 @@ class DocumentIndexInstanceNode(IndexInstanceNode):
         proxy = True
         verbose_name = _('Document index node instance')
         verbose_name_plural = _('Document indexes node instances')
+
+
+class IndexInstanceNodeSearchResult(IndexInstanceNode):
+    class Meta:
+        proxy = True
+        verbose_name = _('Index instance node')
+        verbose_name_plural = _('Index instance nodes')